Rename md049.js in preparation for merging with md050.js.
