+++ demo/markdownlint-browser.js
@@ -4401,10 +4401,10 @@ module.exports = {
 
 /***/ }),
 
-/***/ "../lib/md049.js":
-/*!***********************!*\
-  !*** ../lib/md049.js ***!
-  \***********************/
+/***/ "../lib/md049-md050.js":
+/*!*****************************!*\
+  !*** ../lib/md049-md050.js ***!
+  \*****************************/
 /***/ ((module, __unused_webpack_exports, __webpack_require__) => {
 
 "use strict";
@@ -4634,7 +4634,7 @@ var rules = [
     __webpack_require__(/*! ./md046 */ "../lib/md046.js"),
     __webpack_require__(/*! ./md047 */ "../lib/md047.js"),
     __webpack_require__(/*! ./md048 */ "../lib/md048.js"),
-    __webpack_require__(/*! ./md049 */ "../lib/md049.js"),
+    __webpack_require__(/*! ./md049-md050 */ "../lib/md049-md050.js"),
     __webpack_require__(/*! ./md050 */ "../lib/md050.js"),
     __webpack_require__(/*! ./md051 */ "../lib/md051.js")
 ];
