+++ lib/rules.js
@@ -49,7 +49,7 @@ const rules = [
   require("./md046"),
   require("./md047"),
   require("./md048"),
-  require("./md049"),
+  require("./md049-md050"),
   require("./md050"),
   require("./md051")
 ];
